Properly block registration via the REST API when the appropriate setting is set
